stm32: make resumableperipheral availalble
