+++ embassy-stm32-wpan/src/wba/platform.rs
@@ -61,11 +61,11 @@ use core::task::Poll;
 use embassy_futures::join::join3;
 use embassy_futures::select::select;
 use embassy_stm32::aes::Aes;
-use embassy_stm32::low_power::ResumablePeripheral;
 use embassy_stm32::mode::{Async, Blocking};
 use embassy_stm32::peripherals::{AES as AesPeriph, PKA as PkaPeriph, RNG};
 use embassy_stm32::pka::Pka;
 use embassy_stm32::rng::Rng;
+use embassy_stm32::suspend::ResumablePeripheral;
 use embassy_sync::blocking_mutex::raw::CriticalSectionRawMutex;
 use embassy_sync::mutex::Mutex;
 use embassy_sync::pipe::Pipe;
