+++ embassy-stm32/src/rng.rs
@@ -445,14 +445,13 @@ impl<'d, T: Instance> Drop for Rng<'d, T> {
     }
 }
 
-impl<'d, T: Instance> crate::low_power::SealedSuspendablePeripheral for Rng<'d, T> {
-    #[cfg(all(feature = "low-power", rng_v1))]
+impl<'d, T: Instance> crate::suspend::SealedSuspendablePeripheral for Rng<'d, T> {
+    #[cfg(rng_v1)]
     type InternalState = Peri<'d, T>;
 
-    #[cfg(all(feature = "low-power", not(rng_v1)))]
+    #[cfg(not(rng_v1))]
     type InternalState = (Peri<'d, T>, RngConfig);
 
-    #[cfg(feature = "low-power")]
     fn suspend(self) -> Self::InternalState {
         #[cfg(not(rng_v1))]
         {
@@ -479,7 +478,6 @@ impl<'d, T: Instance> crate::low_power::SealedSuspendablePeripheral for Rng<'d, 
         }
     }
 
-    #[cfg(feature = "low-power")]
     fn resume(state: Self::InternalState) -> Self {
         #[cfg(not(rng_v1))]
         {
